Add task assignments and required certifications to entities
The Task and GroundCrewMember entities have been enhanced. The Task entity now contains a list of required certifications and an Assignment field. Similarly, a task relationship has been added to the GroundCrewMember entity, allowing tasks to be assigned to each member. This allows for a more comprehensive representation of tasks and member abilities.